+++ src/main/java/roomescape/controller/ReservationController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.CreateReservationRequest;
 import roomescape.domain.Reservation;
+import roomescape.dto.CreateReservationRequest;
 
 @RestController
 @RequestMapping("/reservations")
